settings: moved erp settings to apibase
deriving testing from local made to read the missing
local (.env) file which is missing in CI
